Project-Kylo/requirements-kafka.txt
psycopg2-binary==2.9.10
kafka-python==2.2.15
pydantic==2.11.7
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.179.0
aiokafka==0.10.0
PyYAML==6.0.2
